moved reducer logic to src/store/reducers
