+++ src/components/organism/Authentication/Register.tsx
@@ -4,7 +4,7 @@ import Authentication, { AuthType, IOnRegister } from './Authentication';
 
 import { Perform } from '../../../store/reducers/auth';
 
-import { AuthCtx } from '../../../contexts/auth';
+import { AuthCtx } from '../../../store/contexts/auth';
 
 const Register: React.FC = props => {
   const { dispatch } = React.useContext(AuthCtx);
